+++ src/Schema/ISchemaItem.cs
@@ -4,8 +4,9 @@ namespace EditorConfig
 {
     public interface ISchemaItem
     {
-        string Description { get; set; }
-        ImageMoniker Moniker { get; set; }
-        string Name { get; set; }
+        string Name { get; }
+        string Description { get; }
+        ImageMoniker Moniker { get; }
+        bool IsSupported { get; }
     }
 }