+++ src/Schema/SchemaCatalog.cs
@@ -9,63 +9,60 @@ namespace EditorConfig
     {
         public const string Root = "root";
 
-        public static IEnumerable<Property> Properties { get; } = GetProperties();
+        public static IEnumerable<Keyword> Properties { get; } = GetProperties();
         public static IEnumerable<Severity> Severities { get; } = GetSeverities();
 
-        private static IEnumerable<Property> GetProperties()
+        private static IEnumerable<Keyword> GetProperties()
         {
-            return new List<Property>
-            {
-                // Standard properties
-                {new Property(Root, true, Resources.Text.KeywordRoot, KnownMonikers.Property, "true", "false")},
-                {new Property("charset", true, Resources.Text.KeywordCharset, KnownMonikers.Property, "latin1", "utf-8", "utf-8-bom", "utf-16be", "utf-16le", "utf-8-bom")},
-                {new Property("end_of_line", true, Resources.Text.KeywordEndOfLine, KnownMonikers.Property, "lf", "crlf", "cr") },
-                {new Property("indent_style", true, Resources.Text.KeywordIndentStyle, KnownMonikers.Property, "tab", "space")},
-                {new Property("indent_size", true, Resources.Text.KeywordIndentSize, KnownMonikers.Property, "tab") },
-                {new Property("tab_width", true, Resources.Text.KeywordTabWidth, KnownMonikers.Property) },
-                {new Property("trim_trailing_whitespace", false, Resources.Text.KeywordTrimTrailingWhitespace, KnownMonikers.Property, "true", "false")},
-                {new Property("insert_final_newline", false, Resources.Text.KeywordInsertFinalNewline, KnownMonikers.Property, "true", "false")},
+            Category category = Category.Standard;
+            yield return new Keyword(Root, Schema.Text.KeywordRoot, category, "true", "false");
+            yield return new Keyword("charset", Schema.Text.KeywordCharset, category, "latin1", "utf-8", "utf-8-bom", "utf-16be", "utf-16le", "utf-8-bom");
+            yield return new Keyword("end_of_line", Schema.Text.KeywordEndOfLine, category, "lf", "crlf", "cr");
+            yield return new Keyword("indent_style", Schema.Text.KeywordIndentStyle, category, "tab", "space");
+            yield return new Keyword("indent_size", Schema.Text.KeywordIndentSize, category, "tab");
+            yield return new Keyword("tab_width", Schema.Text.KeywordTabWidth, category);
+            yield return new Keyword("trim_trailing_whitespace", Schema.Text.KeywordTrimTrailingWhitespace, category, "true", "false") { IsSupported = false };
+            yield return new Keyword("insert_final_newline", Schema.Text.KeywordInsertFinalNewline, category, "true", "false") { IsSupported = false };
 
-                // C# code analysis
-                {new Property("csharp_style_conditional_delegate_call", true, "Prefer conditional delegate calls.", KnownMonikers.CSFileNode, "true", "false")},
-                {new Property("csharp_style_expression_bodied_accessors", true, "Prefer expression-bodied members.", KnownMonikers.CSFileNode, "true", "false")},
-                {new Property("csharp_style_expression_bodied_constructors", true, "Prefer expression-bodied members", KnownMonikers.CSFileNode, "true", "false")},
-                {new Property("csharp_style_expression_bodied_indexers", true, "Prefer expression-bodied members", KnownMonikers.CSFileNode, "true", "false")},
-                {new Property("csharp_style_expression_bodied_methods", true, "Prefer expression-bodied members.", KnownMonikers.CSFileNode, "true", "false")},
-                {new Property("csharp_style_expression_bodied_operators", true, "Prefer expression-bodied members.", KnownMonikers.CSFileNode, "true", "false")},
-                {new Property("csharp_style_expression_bodied_properties", true, "Prefer expression-bodied members", KnownMonikers.CSFileNode, "true", "false")},
-                {new Property("csharp_style_inlined_variable_declaration", true, "Prefer to inline variable declaration.", KnownMonikers.CSFileNode, "true", "false")},
-                {new Property("csharp_style_pattern_matching_over_as_with_null_check", true, "Prefer pattern matching.", KnownMonikers.CSFileNode, "true", "false")},
-                {new Property("csharp_style_pattern_matching_over_is_with_cast_check", true, "Prefer pattern matching.", KnownMonikers.CSFileNode, "true", "false")},
-                {new Property("csharp_style_throw_expression", true, "Prefer throw expressions", KnownMonikers.CSFileNode, "true", "false")},
-                {new Property("csharp_style_var_elsewhere", true, "Prefer var elsewhere.", KnownMonikers.CSFileNode, "true", "false")},
-                {new Property("csharp_style_var_for_built_in_types", true, "Prefer var for built-in types.", KnownMonikers.CSFileNode, "true", "false")},
-                {new Property("csharp_style_var_for_locals", true, "Prefer var for locals.", KnownMonikers.CSFileNode, "true", "false")},
-                {new Property("csharp_style_var_when_type_is_apparent", true, "Prefer var when type is apparent.", KnownMonikers.CSFileNode, "true", "false")},
+            category = Category.CSharp;
+            yield return new Keyword("csharp_style_conditional_delegate_call", "Prefer conditional delegate calls.", category, "true", "false");
+            yield return new Keyword("csharp_style_expression_bodied_accessors", "Prefer expression-bodied members.", category, "true", "false");
+            yield return new Keyword("csharp_style_expression_bodied_constructors", "Prefer expression-bodied members", category, "true", "false");
+            yield return new Keyword("csharp_style_expression_bodied_indexers", "Prefer expression-bodied members", category, "true", "false");
+            yield return new Keyword("csharp_style_expression_bodied_methods", "Prefer expression-bodied members.", category, "true", "false");
+            yield return new Keyword("csharp_style_expression_bodied_operators", "Prefer expression-bodied members.", category, "true", "false");
+            yield return new Keyword("csharp_style_expression_bodied_properties", "Prefer expression-bodied members", category, "true", "false");
+            yield return new Keyword("csharp_style_inlined_variable_declaration", "Prefer to inline variable declaration.", category, "true", "false");
+            yield return new Keyword("csharp_style_pattern_matching_over_as_with_null_check", "Prefer pattern matching.", category, "true", "false");
+            yield return new Keyword("csharp_style_pattern_matching_over_is_with_cast_check", "Prefer pattern matching.", category, "true", "false");
+            yield return new Keyword("csharp_style_throw_expression", "Prefer throw expressions", category, "true", "false");
+            yield return new Keyword("csharp_style_var_elsewhere", "Prefer var elsewhere.", category, "true", "false");
+            yield return new Keyword("csharp_style_var_for_built_in_types", "Prefer var for built-in types.", category, "true", "false");
+            yield return new Keyword("csharp_style_var_for_locals", "Prefer var for locals.", category, "true", "false");
+            yield return new Keyword("csharp_style_var_when_type_is_apparent", "Prefer var when type is apparent.", category, "true", "false");
 
-                // .NET style analysis
-                {new Property("dotnet_style_coalesce_expression", true, "Prefer null coalescing operator (??).", KnownMonikers.DotNET, "true", "false")},
-                {new Property("dotnet_style_collection_initializer", true, "Prefer collection intiailizers.", KnownMonikers.DotNET, "true", "false")},
-                {new Property("dotnet_style_null_propagation", true, "Prefer null propagation (?.).", KnownMonikers.DotNET, "true", "false")},
-                {new Property("dotnet_style_object_initializer", true, "Prefer object initializers.", KnownMonikers.DotNET, "true", "false")},
-                {new Property("dotnet_style_predefined_type_for_locals_parameters_members", true, "Prefer language types (e.g. int, string, float).", KnownMonikers.DotNET, "true", "false")},
-                {new Property("dotnet_style_predefined_type_for_member_access", true, "Prefer language types (e.g. int, string, float).", KnownMonikers.DotNET, "true", "false")},
-                {new Property("dotnet_style_qualification_for_event", true, "Prefer this for events.", KnownMonikers.DotNET, "true", "false")},
-                {new Property("dotnet_style_qualification_for_field", true, "Prefer this for fields.", KnownMonikers.DotNET, "true", "false")},
-                {new Property("dotnet_style_qualification_for_method", true, "Prefer this for methods.", KnownMonikers.DotNET, "true", "false")},
-                {new Property("dotnet_style_qualification_for_property", true, "Prefer this for properties.", KnownMonikers.DotNET, "true", "false")},
-            };
+            category = Category.DotNet;
+            yield return new Keyword("dotnet_style_coalesce_expression", "Prefer null coalescing operator (??).", category, "true", "false");
+            yield return new Keyword("dotnet_style_collection_initializer", "Prefer collection intiailizers.", category, "true", "false");
+            yield return new Keyword("dotnet_style_null_propagation", "Prefer null propagation (?.).", category, "true", "false");
+            yield return new Keyword("dotnet_style_object_initializer", "Prefer object initializers.", category, "true", "false");
+            yield return new Keyword("dotnet_style_predefined_type_for_locals_parameters_members", "Prefer language types (e.g. int, string, float).", category, "true", "false");
+            yield return new Keyword("dotnet_style_predefined_type_for_member_access", "Prefer language types (e.g. int, string, float).", category, "true", "false");
+            yield return new Keyword("dotnet_style_qualification_for_event", "Prefer this for events.", category, "true", "false");
+            yield return new Keyword("dotnet_style_qualification_for_field", "Prefer this for fields.", category, "true", "false");
+            yield return new Keyword("dotnet_style_qualification_for_method", "Prefer this for methods.", category, "true", "false");
+            yield return new Keyword("dotnet_style_qualification_for_property", "Prefer this for properties.", category, "true", "false");
         }
 
         private static IEnumerable<Severity> GetSeverities()
         {
-            yield return new Severity("none", Resources.Text.SeverityNone, KnownMonikers.None);
-            yield return new Severity("suggestion", Resources.Text.SeveritySuggestion, KnownMonikers.StatusInformation);
-            yield return new Severity("warning", Resources.Text.SeverityWarning, KnownMonikers.StatusWarning);
-            yield return new Severity("error", Resources.Text.SeverityError, KnownMonikers.StatusError);
+            yield return new Severity("none", Schema.Text.SeverityNone, KnownMonikers.None);
+            yield return new Severity("suggestion", Schema.Text.SeveritySuggestion, KnownMonikers.StatusInformation);
+            yield return new Severity("warning", Schema.Text.SeverityWarning, KnownMonikers.StatusWarning);
+            yield return new Severity("error", Schema.Text.SeverityError, KnownMonikers.StatusError);
         }
 
-        public static bool TryGetProperty(string name, out Property property)
+        public static bool TryGetProperty(string name, out Keyword property)
         {
             property = Properties.FirstOrDefault(c => c.Name.Equals(name, StringComparison.OrdinalIgnoreCase));
             return property != null;