+++ src/Schema/Value.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using Microsoft.VisualStudio.Imaging;
+using Microsoft.VisualStudio.Imaging.Interop;
+
+namespace EditorConfig
+{
+    public class Value : ISchemaItem
+    {
+        public Value(string name)
+        {
+            Name = name;
+        }
+
+        public string Name { get; }
+        public string Description { get; }
+        public ImageMoniker Moniker => KnownMonikers.EnumerationItemPublic;
+        public bool IsSupported => true;
+    }
+}