+++ src/SuggestedActions/SuggestedActionsSource.cs
@@ -10,7 +10,7 @@ namespace EditorConfig
 {
     class SuggestedActionsSource : ISuggestedActionsSource
     {
-        private ParseItem _section;
+        private Section _section;
         private EditorConfigDocument _document;
 
         public SuggestedActionsSource(ITextBuffer buffer)
@@ -22,7 +22,7 @@ namespace EditorConfig
         {
             return await Task.Factory.StartNew(() =>
             {
-                _section = _document.ItemsInSpan(range).FirstOrDefault(i => i.ItemType == ItemType.Section);
+                _section = _document.Sections.FirstOrDefault(s => s.Item.Span.Contains(range));
 
                 return _section != null;
             });
@@ -32,7 +32,7 @@ namespace EditorConfig
         {
             if (_section != null)
             {
-                var deleteSection = new DeleteSectionAction(range.Snapshot.TextBuffer, _section.SpanIncludingChildren());
+                var deleteSection = new DeleteSectionAction(range.Snapshot.TextBuffer, _section.Span);
                 yield return new SuggestedActionSet(new[] { deleteSection });
             }
         }