+++ src/Taggers/ErrorTagger.cs
@@ -23,7 +23,7 @@ namespace EditorConfig
             _errorlist = errorlist;
             _textDocument = document;
             _document = EditorConfigDocument.FromTextBuffer(view.TextBuffer);
-            _document.Parsed += DocumentParsed;
+            _document.Validated += DocumentValidated;
 
             ThreadHelper.Generic.BeginInvoke(DispatcherPriority.ApplicationIdle, () =>
             {
@@ -33,7 +33,7 @@ namespace EditorConfig
             });
         }
 
-        private void DocumentParsed(object sender, EventArgs e)
+        private void DocumentValidated(object sender, EventArgs e)
         {
             var span = new SnapshotSpan(_view.TextBuffer.CurrentSnapshot, 0, _view.TextBuffer.CurrentSnapshot.Length);
             TagsChanged?.Invoke(this, new SnapshotSpanEventArgs(span));
@@ -43,7 +43,7 @@ namespace EditorConfig
         {
             var tags = new List<ITagSpan<IErrorTag>>();
 
-            if (_document.IsParsing | !_hasLoaded || !spans.Any() || spans[0].IsEmpty)
+            if (_document.IsParsing || !_hasLoaded || !spans.Any() || spans[0].IsEmpty)
                 return tags;
 
             var line = spans[0].Start.GetContainingLine();